gl_mxmod
gl_mxmod GL ES shader update ./ plane texture

+++ libmx/gl_mxmod/data/tri.vert
@@ -1,5 +1,6 @@
 #version 300 es
-precision mediump float; 
+precision highp float;
+
 layout(location = 0) in vec3 position;   
 layout(location = 1) in vec3 normal;     
 layout(location = 2) in vec2 texCoord;   
@@ -16,6 +17,7 @@ void main() {
     vec4 viewPos4 = view * model * vec4(position, 1.0);
     fragPos = viewPos4.xyz;
     fragNormal = mat3(transpose(inverse(view * model))) * normal;
+    //fragNormal = mat3(view * model) * normal;
     fragTexCoord = texCoord;                                 
     gl_Position = projection * view * vec4(fragPos, 1.0);    
 }
